feat(nextjs): 👻 App Providers | Wagmi & Rainbow
- Minor changes were made to ensure the providers could be included without the need of the layout.
- The console layout was extracted and properly implemented.
- The SignUp page now implements these providers.
