+++ packages/nextjs/app/console/_components/Layout.tsx
@@ -1,0 +1,30 @@
+"use client";
+
+import { useEffect } from "react";
+import { Toaster } from "react-hot-toast";
+import { Footer } from "~~/components/Footer";
+import { Header } from "~~/components/Header";
+import { useNativeCurrencyPrice } from "~~/hooks/scaffold-eth";
+import { useGlobalState } from "~~/services/store/store";
+
+export const ConsoleLayout = ({ children }: { children: React.ReactNode }) => {
+  const price = useNativeCurrencyPrice();
+  const setNativeCurrencyPrice = useGlobalState(state => state.setNativeCurrencyPrice);
+
+  useEffect(() => {
+    if (price > 0) {
+      setNativeCurrencyPrice(price);
+    }
+  }, [setNativeCurrencyPrice, price]);
+
+  return (
+    <>
+      <div className="flex flex-col min-h-screen">
+        <Header />
+        <main className="relative flex flex-col flex-1">{children}</main>
+        <Footer />
+      </div>
+      <Toaster />
+    </>
+  );
+};
