+++ packages/nextjs/app/sign-up/layout.tsx
@@ -1,13 +1,8 @@
 import "@rainbow-me/rainbowkit/styles.css";
-import { ConsoleLayout } from "~console/_components";
 import { AppProviders } from "~~/components/AppProviders";
 
 const Layout = ({ children }: { children: React.ReactNode }) => {
-  return (
-    <AppProviders>
-      <ConsoleLayout>{children}</ConsoleLayout>
-    </AppProviders>
-  );
+  return <AppProviders>{children}</AppProviders>;
 };
 
 export default Layout;
